Remove projections of current vector immediately
If we take the just normalized vector v, and calculate and substract all projections onto v from the remaining vectors, v will not have to be loaded into cache at later times (which was done in the last version of the algorithm). This should improve cache locality of the algorithm (which, however, was not tested).